feat: software ideas
