+++ content/example/feeds/sia.xml
@@ -9,7 +9,7 @@
             <link>https://blog.sia.tech?source=rss----4cd2f4a04096---4</link>
         </image>
         <generator>Medium</generator>
-        <lastBuildDate>Fri, 21 Jul 2023 16:14:06 GMT</lastBuildDate>
+        <lastBuildDate>Fri, 21 Jul 2023 16:59:21 GMT</lastBuildDate>
         <atom:link href="https://blog.sia.tech/feed" rel="self" type="application/rss+xml"/>
         <webMaster><![CDATA[yourfriends@medium.com]]></webMaster>
         <atom:link href="http://medium.superfeedr.com" rel="hub"/>
